# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "RentMe API",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "description": "Value: Bearer ",
      "in": "header",
      "scheme": "bearer"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
    "/api/services/": {
      "post": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "get": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
...
    "/api/dev/": {
      "get": {
        "tags": [
          "Dev"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/dev/upload-images": {
      "post": {
        "tags": [
          "Dev"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/dev/upload-image": {
      "post": {
        "tags": [
          "Dev"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/dev/users": {
      "delete": {
        "tags": [
          "Dev"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}